requirements.txt
Flask
line-bot-sdk
pandas
gunicorn